Fix/suppress various compiler warnings.

+++ src/calendon/aspect-ratio.c
@@ -23,8 +23,8 @@ CnTransform2 cnAspectRatio_fit(CnDimension2u32 desired, CnDimension2u32 actual)
 	CN_ASSERT(0.0f < scale, "Scale must be greater than zero.");
 	CN_ASSERT(scale * desired.width <= actual.width, "Incorrect scale chosen based on width.");
 	CN_ASSERT(scale * desired.height <= actual.height, "Incorrect scale chosen based on height.");
-	CN_ASSERT(abs(actual.width - scale * desired.width) < 2
-		|| abs(actual.height - scale * desired.height) < 2,
+	CN_ASSERT(abs((float)actual.width - scale * (float)desired.width) < 2.0
+		|| abs((float)actual.height - scale * (float)desired.height) < 2.0,
 		"Did not fit the desired resolution into either actual dimension.");
 
 	return cnTransform2_Combine(
